feat: transfer static collections 2 separate class
