+++ src/main/java/UserServlet.java
@@ -1,23 +1,9 @@
 import java.io.IOException;
 
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class UserServlet extends HttpServlet {
-  private static Manager<User> users;
-
-  public UserServlet() {
-    super();
-    users = new Manager<User>();
-    createTestUsers();
-  }
-
-  private void createTestUsers() {
-    users.create(new User("Eugene", "Sokolov"));
-    users.create(new User("John", "Doe"));
-  }
-
+public class UserServlet extends GlobalServlet {
   @Override
   protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
     int id = Integer.parseInt(request.getParameter("id"));
@@ -33,3 +19,4 @@ public class UserServlet extends HttpServlet {
     users.delete(id);
   }
 }
+
